backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-pymongo==2.3.0
pymongo==3.12.0
pyjwt==2.1.0
werkzeug==2.0.1